task5

+++ 0x1A-hash_tables/5-hash_table_print.c
@@ -33,11 +33,11 @@ void print_linked_l(hash_node_t *head, int *prev)
 */
 void hash_table_print(const hash_table_t *ht)
 {
-	int i = 0;
+	unsigned long int i = 0;
 	int prev = 0;
 
 	printf("{");
-	while (i < (int)ht->size)
+	while (i < ht->size)
 	{
 		print_linked_l(ht->array[i], &prev);
 		i++;
